requirements.txt
numpy
faiss-cpu
sentence-transformers
tqdm
python-dotenv
pypdf
google-genai
fastapi
uvicorn
